separated isTrending function and created util
